% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igb_quickload.R
\name{getDefaultAnnotationsUrl}
\alias{getDefaultAnnotationsUrl}
\title{Get name(s) of file(s) IGB loads by default when a user opens a genome.}
\usage{
getDefaultAnnotationsUrl(
  genome_version,
  quickload_root = "http://igbquickload.org/quickload"
)
}
\arguments{
\item{u}{Url of the the genome directory}
}
\description{
Typically there is only one such file, and it usually contains curated,
canonical gene models for a genome. It is usually BED-Detail format.
}
\details{
Note that the file name could be relative to the genome version
root or it could be an absolute URL.

If no file is loaded by default, returns NA.
}
