% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneLengthFunctions.R
\name{getGeneLengthsForPanel}
\alias{getGeneLengthsForPanel}
\title{Get gene model data for species panel}
\usage{
getGeneLengthsForPanel()
}
\value{
Data frame with gene model data for all species in the panel. Includes columns with species (a factor) and log10-transformed gene lengths.
}
\description{
Used in BINF 3121 Markdown "Compare gene lengths across species"
}
