% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneLengthFunctions.R
\name{getGeneModelUrls}
\alias{getGeneModelUrls}
\title{Get gene model annotation BED-detail file URLs for species panel.}
\usage{
getGeneModelUrls()
}
\value{
Character vector with gene model BED-detail file URLs
}
\description{
For BINF 3121 Markdown "Compare gene lengths across species."
Returned species include: H_sapiens (human), M_musculus (mouse),
D_melanogaster (fruit fly), X_tropicalis (frog), S_cerevisiae (yeast),
A_thaliana (mouse-ear cress), O_sativa (rice), P_trichocarpa (cottonwood),
D_rerio (zebrafish)
}
