% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneLengthFunctions.R
\name{getGenomeSizesForPanel}
\alias{getGenomeSizesForPanel}
\title{Create a named vector with species panel genome sizes.}
\usage{
getGenomeSizesForPanel(dat = NULL)
}
\arguments{
\item{dat}{- Output from getGenomeStructuresForAllSpecies. Is optional. If not provided uses output from getGenomeStructuresForPanel.}
}
\value{
Named numeric vector with genome sizes
}
\description{
Used in BINF 3121 Markdown "Compare gene lengths across species."
}
