% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneLengthFunctions.R
\name{getGenomeStructure}
\alias{getGenomeStructure}
\title{Load chromosomes and their sizes for a genome}
\usage{
getGenomeStructure(u)
}
\arguments{
\item{u}{URL path to the genome version directory in IGB quickload}
}
\value{
data frame
}
\description{
This function reads the genome.txt file from an IGB Quickload site.
It returns a data frame with three columns:
}
\details{
species - IGB genome directory (e.g., H_sapiens)
chr - chromosome names (column 1 of genome.txt)
size - chromosome sizes
}
