% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneLengthFunctions.R
\name{getGenomeStructureForOneSpecies}
\alias{getGenomeStructureForOneSpecies}
\title{Load genome panel chromosomes and sizes.}
\usage{
getGenomeStructureForOneSpecies(u)
}
\arguments{
\item{u}{URL of gene model annotation file}
}
\value{
Data frame with chromosome names, sizes, and species (a factor)
}
\description{
Used for BINF 3121 Markdown "Compare gene lengths across species."
Reads the genome.txt files from an IGB Quickload site.
Removes poorly-assembled or alternative chromosomes.
}
