% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneLengthFunctions.R
\name{getGenomeStructuresForPanel}
\alias{getGenomeStructuresForPanel}
\title{Load chromosomes and their sizes for all genomes in the panel.}
\usage{
getGenomeStructuresForPanel()
}
\value{
Data frame with chromosome names, sizes, and species (a factor)
}
\description{
Used in BINF 3121 Markdown "Compare gene lengths across species."
Uses getGenomeStructureForOneSpecies.
}
