% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igb_quickload.R
\name{getGenomeVersionUrls}
\alias{getGenomeVersionUrls}
\title{Get data frame with genome version directories from an IGB Quickload site}
\usage{
getGenomeVersionUrls(url = "http://igbquickload.org/quickload")
}
\arguments{
\item{u}{Url of the IGB Quickload site}
}
\value{
Data frame with contents.txt
}
\description{
This function loads the contents.txt file from an IGB Quickload site.
}
