% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneLengthFunctions.R
\name{getMedianGeneLengthsForPanel}
\alias{getMedianGeneLengthsForPanel}
\title{Get gene model median lengths for species panel}
\usage{
getMedianGeneLengthsForPanel(dat)
}
\arguments{
\item{dat}{- Output of getGeneLengthsForPanel}
}
\value{
Named numeric vector with median gene lengths, sorted by size
}
\description{
Used in BINF 3121 Markdown "Compare gene lengths across species"
}
