% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneLengthFunctions.R
\name{getDataForOneSpecies}
\alias{getDataForOneSpecies}
\title{Get gene model data for a species.}
\usage{
getDataForOneSpecies(url)
}
\arguments{
\item{url}{URL of species BED-detail file to read}
}
\value{
Data frame with gene model data, including species (as a factor) and log10-transformed gene lengths.
}
\description{
For BINF 3121 Markdown "Compare gene lengths across species."
}
