% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneLengthFunctions.R
\name{getGenomeSizesForPanel}
\alias{getGenomeSizesForPanel}
\title{Creates a named vector with species panel genome sizes}
\usage{
getGenomeSizesForPanel(dat = NULL)
}
\arguments{
\item{dat}{- Output from getGenomeStructuresForAllSpecies (optional)}
}
\value{
Named numeric vector with genome sizes
}
\description{
Used in BINF 3121 Markdown "Compare gene lengths across species"
}
